+++ CHANGELOG.rst
@@ -12,6 +12,10 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
 * Use case counting all samples of a project and the subset of samples having been received at the lab
 
+* Shows the number of received sample to each customer (`#36 <https://github.com/qbicsoftware/sample-tracking-status-overview/pull/36>`_)
+
+* Add resource service for project sample status counts. (`#34 <https://github.com/qbicsoftware/sample-tracking-status-overview/pull/34>`_)
+
 * Interface for counting samples and received samples given a project code
 
 * Add resource service for project samples status list. (`#34 <https://github.com/qbicsoftware/sample-tracking-status-overview/pull/34>`_)